api/auth_service/requirements.txt
flask
flasgger
flask-sqlalchemy
passlib
pyjwt
python-dotenv
pymysql